LayoutTests/accessibility/ios-simulator/fieldset-traits-expected.txt

Personalia:
Name:
Email:
This tests that elements in a fieldset are having the correct traits.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputname.hasContainedByFieldsetTrait is true
PASS fieldset.isEqual(inputname.fieldsetAncestorElement()) is true
PASS successfullyParsed is true

TEST COMPLETE
